Changes to mongodb connection
